+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Add a custom drawer for `ValueListener` [[#41](https://github.com/DarkRewar/BaseTool/issues/41)]
+- Add array and range extensions [[#39](https://github.com/DarkRewar/BaseTool/issues/39)]
 
 ### Changes